# PERF: fastpath DataFrame constructor from BlockManager