--- a/pandas/core/apply.py
+++ b/pandas/core/apply.py
@@ -166,9 +166,7 @@ class FrameApply(metaclass=abc.ABCMeta):
         elif isinstance(self.f, np.ufunc):
             with np.errstate(all="ignore"):
                 results = self.obj._mgr.apply("apply", func=self.f)
-            return self.obj._constructor(
-                data=results, index=self.index, columns=self.columns, copy=False
-            )
+            return type(self.obj)._from_mgr(results)
 
         # broadcasting
         if self.result_type == "broadcast":