--- a/pandas/core/generic.py
+++ b/pandas/core/generic.py
@@ -230,15 +230,18 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
                 mgr = mgr.astype(dtype=dtype)
         return mgr
 
-    @classmethod
-    def _from_mgr(cls: Type[FrameOrSeries], mgr: BlockManager) -> FrameOrSeries:
+    def _from_mgr(self: FrameOrSeries, mgr: BlockManager) -> FrameOrSeries:
         """
         Fastpath constructor for if we have only a BlockManager.
         """
-        constructor = cls._constructor.fget(cls)  # type: ignore
-        obj = object.__new__(constructor)
-        NDFrame.__init__(obj, mgr)
-        return obj
+        constr = self._constructor
+        if constr is type(self):
+            # See GH#33357 for why this check is needed
+            obj = object.__new__(constr)
+            NDFrame.__init__(obj, mgr)
+            return obj
+
+        return constr(mgr)
 
     # ----------------------------------------------------------------------
 
@@ -1369,7 +1372,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
             return self
 
         new_data = self._mgr.apply(operator.invert)
-        result = type(self)._from_mgr(new_data)
+        result = self._from_mgr(new_data)
         return result.__finalize__(self, method="__invert__")
 
     def __nonzero__(self):
@@ -3375,7 +3378,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         new_data = self._mgr.take(
             indices, axis=self._get_block_manager_axis(axis), verify=True
         )
-        return type(self)._from_mgr(new_data).__finalize__(self, method="take")
+        return self._from_mgr(new_data).__finalize__(self, method="take")
 
     def _take_with_is_copy(self: FrameOrSeries, indices, axis=0) -> FrameOrSeries:
         """
@@ -3581,7 +3584,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         """
         assert isinstance(slobj, slice), type(slobj)
         axis = self._get_block_manager_axis(axis)
-        result = type(self)._from_mgr(self._mgr.get_slice(slobj, axis=axis))
+        result = self._from_mgr(self._mgr.get_slice(slobj, axis=axis))
         result = result.__finalize__(self)
 
         # this could be a view
@@ -4518,7 +4521,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         if copy and new_data is self._mgr:
             new_data = new_data.copy()
 
-        return type(self)._from_mgr(new_data).__finalize__(self)
+        return self._from_mgr(new_data).__finalize__(self)
 
     def filter(
         self: FrameOrSeries,
@@ -5284,7 +5287,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         else:
             f = lambda: self._mgr.consolidate()
             cons_data = self._protect_consolidate(f)
-            return type(self)._from_mgr(cons_data).__finalize__(self)
+            return self._from_mgr(cons_data).__finalize__(self)
 
     @property
     def _is_mixed_type(self) -> bool_t:
@@ -5313,10 +5316,10 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         return True
 
     def _get_numeric_data(self):
-        return type(self)._from_mgr(self._mgr.get_numeric_data()).__finalize__(self)
+        return self._from_mgr(self._mgr.get_numeric_data()).__finalize__(self)
 
     def _get_bool_data(self):
-        return type(self)._from_mgr(self._mgr.get_bool_data()).__finalize__(self)
+        return self._from_mgr(self._mgr.get_bool_data()).__finalize__(self)
 
     # ----------------------------------------------------------------------
     # Internal Interface Methods
@@ -5583,7 +5586,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         else:
             # else, only a single dtype is given
             new_data = self._mgr.astype(dtype=dtype, copy=copy, errors=errors,)
-            return type(self)._from_mgr(new_data).__finalize__(self, method="astype")
+            return self._from_mgr(new_data).__finalize__(self, method="astype")
 
         # GH 19920: retain column metadata after concat
         result = pd.concat(results, axis=1, copy=False)
@@ -5697,7 +5700,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         """
         data = self._mgr.copy(deep=deep)
         self._clear_item_cache()
-        return type(self)._from_mgr(data).__finalize__(self, method="copy")
+        return self._from_mgr(data).__finalize__(self, method="copy")
 
     def __copy__(self: FrameOrSeries, deep: bool_t = True) -> FrameOrSeries:
         return self.copy(deep=deep)
@@ -5748,19 +5751,15 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         validate_bool_kwarg(timedelta, "timedelta")
         validate_bool_kwarg(coerce, "coerce")
         validate_bool_kwarg(copy, "copy")
-        return (
-            type(self)
-            ._from_mgr(
-                self._mgr.convert(
-                    datetime=datetime,
-                    numeric=numeric,
-                    timedelta=timedelta,
-                    coerce=coerce,
-                    copy=copy,
-                )
+        return self._from_mgr(
+            self._mgr.convert(
+                datetime=datetime,
+                numeric=numeric,
+                timedelta=timedelta,
+                coerce=coerce,
+                copy=copy,
             )
-            .__finalize__(self)
-        )
+        ).__finalize__(self)
 
     def infer_objects(self: FrameOrSeries) -> FrameOrSeries:
         """
@@ -5803,19 +5802,11 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         # numeric=False necessary to only soft convert;
         # python objects will still be converted to
         # native numpy numeric types
-        return (
-            type(self)
-            ._from_mgr(
-                self._mgr.convert(
-                    datetime=True,
-                    numeric=False,
-                    timedelta=True,
-                    coerce=False,
-                    copy=True,
-                )
+        return self._from_mgr(
+            self._mgr.convert(
+                datetime=True, numeric=False, timedelta=True, coerce=False, copy=True,
             )
-            .__finalize__(self, method="infer_objects")
-        )
+        ).__finalize__(self, method="infer_objects")
 
     def convert_dtypes(
         self: FrameOrSeries,
@@ -6918,7 +6909,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
             **kwargs,
         )
 
-        result = type(self)._from_mgr(new_data)
+        result = self._from_mgr(new_data)
         if axis == 1:
             result = result.T
         if inplace:
@@ -8533,7 +8524,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
             if copy and fdata is self._mgr:
                 fdata = fdata.copy()
 
-            left = type(self)._from_mgr(fdata)
+            left = self._from_mgr(fdata)
 
             if ridx is None:
                 right = other
@@ -8685,7 +8676,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
             new_data = self._mgr.putmask(
                 mask=cond, new=other, align=align, axis=block_axis,
             )
-            result = type(self)._from_mgr(new_data)
+            result = self._from_mgr(new_data)
             return self._update_inplace(result)
 
         else:
@@ -8697,7 +8688,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
                 try_cast=try_cast,
                 axis=block_axis,
             )
-            result = type(self)._from_mgr(new_data)
+            result = self._from_mgr(new_data)
             return result.__finalize__(self)
 
     _shared_docs[
@@ -8970,7 +8961,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         else:
             return self.tshift(periods, freq)
 
-        return type(self)._from_mgr(new_data).__finalize__(self, method="shift")
+        return self._from_mgr(new_data).__finalize__(self, method="shift")
 
     def slice_shift(self: FrameOrSeries, periods: int = 1, axis=0) -> FrameOrSeries:
         """
@@ -11224,7 +11215,7 @@ def _make_cum_function(
 
         result = self._mgr.apply(block_accum_func)
 
-        obj = type(self)._from_mgr(result)
+        obj = self._from_mgr(result)
         return obj.__finalize__(self, method=name)
 
     return set_function_name(cum_func, name, cls)