@jbrockmendel 2020-04-07
> makes sense but would changing obj._constructor -> type(obj)._from_mgr create issues with subclassing.

Fair point.  If _constructor is something other than `type(self)` this would be change behavior.  Will update.

> I recall a recent discussion that _constructor should be type(self)

This sounds familiar.  I'd be +1.