@jorisvandenbossche 2020-04-07
>  Moreover, since users shouldn't be passing BlockManagers around anyway, we might be able to get that case out of the DataFrame constructor entirely.

This is however used by downstream libraries (eg pyarrow).

> I recall a recent discussion that _constructor should be type(self)

That is currently not required to be the case (and for geopandas it also is not the case). After that earlier discussion, this issue was opened about that: https://github.com/pandas-dev/pandas/issues/32638