@simonjayhawkins 2020-04-07
@jorisvandenbossche thanks for the link. you wrote

> we also don't require _constructor to be a class, it can be a constructor function

so https://github.com/pandas-dev/pandas/pull/33357#discussion_r404992724 is probably not applicable.