@jorisvandenbossche 2020-04-07
The only requirement we have is that `_constructor` returns a callable, so `obj._constructor._from_mgr` will not necessarily work.