@jbrockmendel 2020-04-07
> Basically check that data is a BlockManager and that all other keywords are None? (that are some more checks of course, so would need to check if that doesn't undo the gain)

if this or something like it isnt accepted, then that may be worthwhile.  if we do get this fastpath, then id prefer not to make the already-complicated constructor more complicated.

> The only requirement we have is that _constructor returns a callable, so obj._constructor._from_mgr will not necessarily work.

How about

```
def _from_mgr(self, mgr):
    constr = self._constructor
    if constr is type(self):
         [the current implementation]
    else:
        return constr(mgr)
```