@jorisvandenbossche 2020-04-07
> How about

Then I personally think having the shortcut in the main constructor is the simpler solution