@jbrockmendel 2020-04-08
On further reflection, I'm on board with Joris's suggestion.  If at some point we add restrictions to _constructor that allow _from_mgr to be re-simplified, ill revisit at that point.